requirements.txt
cx-Freeze==6.15.12
inquirer==3.1.4
python-dotenv==1.0.0
requests==2.31.0
selenium==4.16.0
